set recsize to SIZE_MAX when use of small records is unnecessary; doing so avoids the cost of invoking SSL_write / ptls_send every 16KB

+++ include/h2o/socket.h
@@ -134,7 +134,7 @@ struct st_h2o_socket_t {
     struct {
         uint8_t state; /* one of H2O_SOCKET_LATENCY_STATE_* */
         uint8_t notsent_is_minimized : 1;
-        uint16_t suggested_tls_payload_size;
+        size_t suggested_tls_payload_size; /* suggested TLS record payload size, or SIZE_MAX when no need to restrict */
         size_t suggested_write_size; /* SIZE_MAX if no need to optimize for latency */
     } _latency_optimization;
 };
